CLDINT-661 - Implement first step of First Login Wizard (Complete your profile) - Finish avatar JS implementation

+++ wizard-extension/config/src/main/java/com/exoplatform/cloudworkspaces/cloudlogin/rest/CloudLoginRestService.java
@@ -18,39 +18,45 @@
 package com.exoplatform.cloudworkspaces.cloudlogin.rest;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 import javax.annotation.security.RolesAllowed;
+import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.CacheControl;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
+import org.apache.commons.fileupload.FileUploadException;
 import org.exoplatform.common.http.HTTPStatus;
 import org.exoplatform.services.log.ExoLogger;
 import org.exoplatform.services.log.Log;
 import org.exoplatform.services.rest.resource.ResourceContainer;
 import org.exoplatform.services.security.ConversationState;
 import org.exoplatform.services.wcm.utils.WCMCoreUtils;
-import org.exoplatform.social.core.identity.model.Profile;
-import org.exoplatform.social.core.image.ImageUtils;
-import org.exoplatform.social.core.model.AvatarAttachment;
-import org.exoplatform.social.webui.Utils;
 import org.exoplatform.upload.UploadResource;
 import org.exoplatform.upload.UploadService;
+
+import com.exoplatform.cloudworkspaces.cloudlogin.CloudLoginService;
 
 @Path(CloudLoginRestService.WS_ROOT_PATH)
 public class CloudLoginRestService implements ResourceContainer {
   
   private static Log logger = ExoLogger.getLogger(CloudLoginRestService.class);
 
+  protected final static String WS_PORTAL_PATH = "/portal/rest";
   protected final static String WS_ROOT_PATH = "/cloudlogin";
-  public final static Long MAX_AVATAR_LENGTH = 2000000L;
+  protected final static String WS_UPLOAD_PATH = "/uploadavatar";
+  protected final static String WS_PROFILE_PATH = "/updateprofile";
+  public final static String WS_UPLOAD_PARAM_UPLOAD_ID = "uploadId";
+  
+  public final static Integer MAX_AVATAR_LENGTH = 2000000;
   public final static String AVATAR_UPLOAD_ID = "cloudloginavatar";
   
   @SuppressWarnings("serial")
@@ -64,10 +70,54 @@ public class CloudLoginRestService implements ResourceContainer {
     }
   };
   
-  protected UploadService uploadService;
+  private UploadService uploadService;
+  private CloudLoginService cloudLoginService;
   
-  public CloudLoginRestService() {
+  public CloudLoginRestService(CloudLoginService cloudLoginService) {
     uploadService = WCMCoreUtils.getService(UploadService.class);
+    this.cloudLoginService = cloudLoginService;
+  }
+  
+  /**
+   * This method is used to upload an image into server, create a temporary thumbnails into JCR and returns URL of this image
+   * 
+   * @param servletRequest
+   * @param uploadId
+   * @return
+   */
+  @POST
+  @Path(CloudLoginRestService.WS_UPLOAD_PATH)
+  @Produces(MediaType.TEXT_PLAIN)
+  @RolesAllowed("users")
+  public Response uploadAvatar(@Context HttpServletRequest servletRequest, @QueryParam(CloudLoginRestService.WS_UPLOAD_PARAM_UPLOAD_ID) String uploadId) {
+
+    CacheControl cacheControl = new CacheControl();
+    String avatarUri = "";
+
+    if(uploadId != null && servletRequest != null) {
+      try {
+        uploadService.addUploadLimit(uploadId, MAX_AVATAR_LENGTH);
+        uploadService.createUploadResource(servletRequest);
+        
+        // Try to create a JCR node to have a thumbnails displayable by client
+        UploadResource upResource = uploadService.getUploadResource(uploadId);
+        
+        // Create Temporary avatar node
+        cloudLoginService.createTempAvatarNode(upResource);
+        
+        // Manage cache
+        cacheControl.setNoCache(true);
+        cacheControl.setNoStore(true);
+      }
+      catch (FileUploadException e) {
+        logger.warn("Cannot upload image with id: " + uploadId + " (" + e.getMessage() + ")");
+      }
+      catch (Exception e) {
+        logger.warn("Cannot upload image with id: " + uploadId + " (" + e.getMessage() + ")");
+      }
+    }
+    
+    return Response.ok(avatarUri).cacheControl(cacheControl).build();
   }
   
   /**
@@ -81,9 +131,10 @@ public class CloudLoginRestService implements ResourceContainer {
    * @throws Exception
    */
   @GET
-  @Path("/setavatar")
+  @Path(CloudLoginRestService.WS_PROFILE_PATH)
+  @Produces(MediaType.TEXT_PLAIN)
   @RolesAllowed("users")
-  public Response createAvatar(
+  public Response updateProfile(
                 @QueryParam("fileName") String fileName,
                 @QueryParam("uploadId") String uploadId) {
 
@@ -91,45 +142,37 @@ public class CloudLoginRestService implements ResourceContainer {
     cacheControl.setNoCache(true);
     
     if(fileName != null && uploadId != null) {
-      String userId = ConversationState.getCurrent().getIdentity().getUserId();
       
       try {
+        String userId = ConversationState.getCurrent().getIdentity().getUserId();
         
         // Get resource uploaded into server
         UploadResource resource = uploadService.getUploadResource(uploadId);
-        File file = new File(resource.getStoreLocation());
-        FileInputStream inputStream = new FileInputStream(file);
-        String mimeType = resource.getMimeType();
-        
-        if (! ACCEPTED_MIME_TYPES.contains(mimeType)) {
-          uploadService.removeUploadResource(uploadId);
-          logger.warn("Mimetype " + mimeType + " is not accepted" + mimeType);
-          return Response.ok("Your file is not in proper format", MediaType.TEXT_PLAIN).cacheControl(cacheControl).build();
-        }
-        
-        if(file.length() >= MAX_AVATAR_LENGTH) {
-          uploadService.removeUploadResource(uploadId);
-          logger.warn("File size is too large: " + file.length());
-          return Response.ok("File size is too large: " + file.length() + " (Max=" + MAX_AVATAR_LENGTH + ")", MediaType.TEXT_PLAIN).cacheControl(cacheControl).build();
-        }
-        
-        // Create avatar attachement
-        AvatarAttachment avatarAttachment = ImageUtils.createResizedAvatarAttachment(inputStream, 200, 0, null, fileName, mimeType, null);
-        if(avatarAttachment == null) {
-          avatarAttachment = new AvatarAttachment(null, fileName, resource.getMimeType(), inputStream, null, System.currentTimeMillis());
-        }
-        
-        // Update profile with new avatar
-        Profile p = Utils.getUserIdentity(userId, true).getProfile();
-        p.setProperty(Profile.AVATAR, avatarAttachment);
-        Map<String, Object> props = p.getProperties();
-        // Removes avatar url and resized avatar
-        for (String key : props.keySet()) {
-          if (key.startsWith(Profile.AVATAR + ImageUtils.KEY_SEPARATOR)) {
-            p.removeProperty(key);
+        if(resource != null) {
+          File file = new File(resource.getStoreLocation());
+          String mimeType = resource.getMimeType();
+          
+          if (! ACCEPTED_MIME_TYPES.contains(mimeType)) {
+            uploadService.removeUploadResource(uploadId);
+            logger.warn("Mimetype " + mimeType + " is not accepted" + mimeType);
+            return Response.ok("Your file is not in proper format", MediaType.TEXT_PLAIN).cacheControl(cacheControl).build();
           }
+          
+          if(file.length() >= MAX_AVATAR_LENGTH) {
+            uploadService.removeUploadResource(uploadId);
+            logger.warn("File size is too large: " + file.length());
+            return Response.ok("File size is too large: " + file.length() + " (Max=" + MAX_AVATAR_LENGTH + ")", MediaType.TEXT_PLAIN).cacheControl(cacheControl).build();
+          }
+          
+          // Update user profile
+          cloudLoginService.updateProfile(userId, resource);
+  
+          // Delete temporary avatar
+          cloudLoginService.deleteTempAvatarNode();
         }
-        Utils.getIdentityManager().updateProfile(p);
+        else {
+          logger.warn("Upload resource with id " + uploadId + " cannot be found.");
+        }
       }
       catch(Exception e) {
         logger.error("WS " + WS_ROOT_PATH + "/setavatar has a problem with getting node of avatar", e);
@@ -141,5 +184,13 @@ public class CloudLoginRestService implements ResourceContainer {
     }
     
     return Response.ok().cacheControl(cacheControl).build();
+  }
+  
+  public static String getUploadWsPath() {
+    return WS_PORTAL_PATH + WS_ROOT_PATH + WS_UPLOAD_PATH;
+  }
+  
+  public static String getProfileWsPath() {
+    return WS_PORTAL_PATH + WS_ROOT_PATH + WS_PROFILE_PATH;
   }
 }
